[#24556] Combine sim settings & plot/export dialogs
