+++ core/resources/l10n/messages.properties
@@ -452,9 +452,10 @@ simedtdlg.but.delete = Delete
 simedtdlg.title.Editsim = Edit simulation
 simedtdlg.title.MultiSimEdit = Multi-simulation edit
 simedtdlg.title.MultiSimEdit.ttip = <html>You are editing the following simulations:<br>
-simedtdlg.lbl.Simname = Simulation name:
+simedtdlg.lbl.Simname = Name:
 simedtdlg.tab.Launchcond = Launch conditions
 simedtdlg.tab.Simopt = Simulation options
+simedtdlg.tab.Settings = Settings
 simedtdlg.tab.Plotdata = Plot data
 simedtdlg.tab.CustomExpressions = Custom expressions
 simedtdlg.tab.Exportdata = Export data
@@ -533,6 +534,10 @@ simedtdlg.IntensityDesc.Medium = Medium
 simedtdlg.IntensityDesc.High = High
 simedtdlg.IntensityDesc.Veryhigh = Very high
 simedtdlg.IntensityDesc.Extreme = Extreme
+
+! SimulationConfigDialog
+SimulationConfigDialog.tab.plotDis.ttip = Plotting not supported for multi-simulation editing
+SimulationConfigDialog.tab.expDis.ttip = Exporting not supported for multi-simulation editing
 
 SimulationExtension.airstart.name.alt = Air-start ({alt})
 SimulationExtension.airstart.name.altvel = Air-start ({alt}, {vel})
