+++ core/resources/l10n/messages_uk_UA.properties
@@ -321,7 +321,7 @@ simedtdlg.but.savedefault = Save as default
 simedtdlg.but.add = Add
 simedtdlg.but.delete = Delete
 simedtdlg.title.Editsim = Edit simulation
-simedtdlg.lbl.Simname = Simulation name:
+simedtdlg.lbl.Simname = Name:
 simedtdlg.tab.Launchcond = Launch conditions
 simedtdlg.tab.Simopt = Simulation options
 simedtdlg.tab.Plotdata = Plot data
